pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.51.0",
    "fpdf2>=2.7.8",
]
